[refactor api] Update transpilation to opt-in only for browser applications using "browser" in package.json.
