+++ .npmignore
@@ -11,9 +11,6 @@ pids
 *.seed
 *.pid.lock
 
-# non-compiled code
-src
-
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
