+++ cli.js
