+++ colorize.js
