+++ logstash.js
