+++ splat.js
