+++ test/label.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const assume = require('assume');
-const label = require('../src/label');
+const label = require('../label');
 const helpers = require('./helpers');
 const { MESSAGE } = require('triple-beam');
