+++ uncolorize.js
